refactor: isolate leaf slack handlers
