+++ apps/leaf/src/bot.ts
@@ -1,110 +1,31 @@
 import { verifyDashboardSession } from "@autumn/auth";
-import { buildCatalogDecisionModel } from "@autumn/render";
-import type { AppEnv } from "@autumn/shared";
 import { createSlackAdapter } from "@chat-adapter/slack";
 import { verifySlackSignature } from "@chat-adapter/slack/webhook";
 import { createPostgresState } from "@chat-adapter/state-pg";
 import { createWebAdapter } from "@chat-adapter/web";
-import type { Attachment, Message, Thread } from "chat";
 import { Chat } from "chat";
-import { answerAgentQuestion } from "./internal/agentRuntime/actions/answerAgentQuestion/answerAgentQuestion.js";
-import { resolveAgentCatalogDecision } from "./internal/agentRuntime/actions/resolveCatalogDecision/resolveAgentCatalogDecision.js";
-import { withdrawAgentTurnApproval } from "./internal/agentRuntime/actions/withdrawAgentTurnApproval/withdrawAgentTurnApproval.js";
-import type { AgentContextMessage } from "./internal/agentRuntime/domain/agentTurnContext.js";
-import { chatApprovalRepo } from "./internal/approvals/repos/chatApprovalRepo.js";
 import { handleApprovalAction } from "./internal/approvals/surfaces/slack/decide.js";
-import {
-	postApprovalCardForRow,
-	presentApproval,
-} from "./internal/approvals/surfaces/slack/present.js";
-import { editSupersededApprovalCards } from "./internal/approvals/surfaces/slack/superseded.js";
 import { handleViewPayloadAction } from "./internal/approvals/surfaces/slack/viewPayload.js";
 import { ensureWebChatAuth } from "./internal/installations/actions/ensureWebChatAuth.js";
-import { getInstallationOAuthAccessToken } from "./internal/installations/actions/getInstallationOAuthAccessToken.js";
 import { handleStopAction } from "./internal/runs/handleStopAction.js";
-import {
-	dispatchThreadMessage,
-	hasQueuedThreadMessage,
-} from "./internal/runs/runCoordinator.js";
-import {
-	type ActiveRun,
-	closeRun,
-	registerRun,
-	runKeyForThread,
-} from "./internal/runs/runRegistry.js";
-import { shouldUseSlackAdminInstallationForWorkspace } from "./internal/slackAdmin/access.js";
 import { decrypt } from "./lib/crypto.js";
-import { db } from "./lib/db.js";
 import { env } from "./lib/env.js";
-import {
-	addLeafContext,
-	createLeafSessionContext,
-	logger as rootLogger,
-} from "./lib/logger.js";
-import { runSlackAgentTurn } from "./providers/slack/actions/runSlackAgentTurn.js";
-import { getSlackWorkspaceId } from "./providers/slack/context.js";
+import { logger as rootLogger } from "./lib/logger.js";
 import {
 	getSlackEventWorkspaceId,
 	normalizeSlackEventsBody,
 } from "./providers/slack/events.js";
-import { createEveSlackPresenter } from "./providers/slack/evePresenter.js";
+import { handleSlackCatalogDecision } from "./providers/slack/handlers/handleSlackCatalogDecision.js";
+import { handleSlackMessage } from "./providers/slack/handlers/handleSlackMessage.js";
+import { handleSlackQuestionAnswer } from "./providers/slack/handlers/handleSlackQuestionAnswer.js";
+import { handleSlackSlashCommand } from "./providers/slack/handlers/handleSlackSlashCommand.js";
+import { findSlackInstallationForWorkspace } from "./providers/slack/installations.js";
 import {
-	fetchSlackAttachmentFallback,
-	getSlackFilesFromRaw,
-} from "./providers/slack/files.js";
-import { findInstallationWithOrg } from "./providers/slack/installations.js";
-import { getRecentMessages } from "./providers/slack/threadContext.js";
-import {
-	ANSWER_QUESTION_ACTION,
-	CATALOG_DECISION_ACTION,
-	type CatalogDecisionButtonPayload,
-	catalogDecisionCard,
-	catalogDecisionSubmittedCard,
-	indexedActionIds,
-	type QuestionButtonPayload,
-	questionAnsweredCard,
-	questionCard,
-} from "./ui/eveCards.js";
-import {
-	GENERIC_FAILURE_MESSAGE,
-	genericFailureWithDetail,
-	NO_REPLY_MESSAGE,
-	POST_FORMATTING_FAILED_MESSAGE,
-	QUESTION_ANSWER_FAILED_MESSAGE,
-	questionAnswerFailedWithDetail,
-	RUN_STOPPED_FOR_TIME_MESSAGE,
-	RUN_TIMED_OUT_MESSAGE,
-	runStoppedByUserNotice,
-} from "./ui/messages.js";
-import type { ReplyTarget } from "./ui/progress.js";
-import { createStatusTicker } from "./ui/statusTicker.js";
+	catalogDecisionActionIds,
+	questionAnswerActionIds,
+} from "./providers/slack/presenters/interactionCards.js";
 
 export const chatAdapterNames = ["slack", "web"];
-
-const getSlackAdminProvider = () =>
-	`slack_admin:${env.SLACK_CLIENT_ID}` as const;
-
-const findSlackInstallationForWorkspace = async ({
-	workspaceId,
-}: {
-	workspaceId: string;
-}) => {
-	if (
-		shouldUseSlackAdminInstallationForWorkspace({
-			configuredWorkspaceId: env.SLACK_ADMIN_WORKSPACE_ID,
-			isProduction: process.env.NODE_ENV === "production",
-			workspaceId,
-		})
-	) {
-		const adminInstallation = await findInstallationWithOrg(
-			getSlackAdminProvider(),
-			workspaceId,
-		);
-		if (adminInstallation) return adminInstallation;
-	}
-
-	return await findInstallationWithOrg("slack", workspaceId);
-};
 
 export const bot = new Chat({
 	userName: env.CHAT_NAME,
@@ -155,17 +76,12 @@ export const bot = new Chat({
 						hasOrg: Boolean(session?.activeOrganizationId),
 					},
 				});
-				if (!session?.activeOrganizationId) {
-					return null;
-				}
-				// Mint/refresh this user's scope-bound MCP OAuth credential at the
-				// cookie boundary, so downstream reads never run unauthenticated.
+				if (!session?.activeOrganizationId) return null;
 				await ensureWebChatAuth({
 					orgId: session.activeOrganizationId,
 					userId: session.userId,
 					userScopes: session.scopes,
 				});
-				// Preserve the legacy web adapter's user-id format for old history.
 				return { id: `${session.userId}~${session.activeOrganizationId}` };
 			},
 		}),
@@ -174,586 +90,22 @@ export const bot = new Chat({
 		keyPrefix: "chat",
 		url: env.CHAT_STATE_DATABASE_URL,
 	}),
-	// Handlers run immediately; the run coordinator serializes new runs per
-	// thread and routes mid-run messages (stop keywords, live follow-ups).
 	concurrency: "concurrent",
 });
 
-// One key per physical Slack thread, shared by message and dispatch paths.
-const slackRunKey = ({
-	channelId,
-	raw,
-	threadId,
-}: {
-	channelId: string;
-	raw: unknown;
-	threadId: string;
-}) =>
-	runKeyForThread({
-		channelId,
-		provider: "slack",
-		threadId,
-		workspaceId: getSlackWorkspaceId(raw),
-	});
-
-const ERROR_NOTICE_MAX = 160;
-
-/** One clean, human line about what failed — never a stack or a shrug. */
-const errorNotice = (error: unknown) => {
-	const message = error instanceof Error ? error.message : String(error);
-	if (/invalid_blocks/i.test(message)) return POST_FORMATTING_FAILED_MESSAGE;
-	if (/timed out|timeout/i.test(message)) return RUN_TIMED_OUT_MESSAGE;
-	const detail = message.replace(/\s+/g, " ").trim().slice(0, ERROR_NOTICE_MAX);
-	return detail ? genericFailureWithDetail(detail) : GENERIC_FAILURE_MESSAGE;
-};
-
-type RunAndReplyInput = {
-	attachments?: Attachment[];
-	/** One-turn structured context (e.g. a submitted catalog decision card). */
-	clientContext?: Record<string, unknown>;
-	channelId: string;
-	providerUserId: string;
-	raw: unknown;
-	/** Reacts to the triggering message (👀 while working, ❌ on failure). */
-	react?: (input: { action: "add" | "remove"; emoji: string }) => Promise<void>;
-	recentMessages?: AgentContextMessage[];
-	runKey: string;
-	target: ReplyTarget;
-	text: string;
-	threadId: string;
-};
-
-const runAndReply = async ({
-	channelId,
-	attachments,
-	clientContext,
-	providerUserId,
-	raw,
-	react,
-	recentMessages: recentMessagesInput,
-	runKey,
-	target,
-	text,
-	threadId,
-}: Omit<RunAndReplyInput, "recentMessages"> & {
-	/** A thunk defers the Slack history fetch to run start (queued messages
-	 * must see turns that finished before them) while still overlapping it
-	 * with the installation lookup. */
-	recentMessages?:
-		| AgentContextMessage[]
-		| (() => Promise<AgentContextMessage[]>);
-}) => {
-	let logger = rootLogger;
-	let run: ActiveRun | undefined;
-	const ticker = createStatusTicker(target);
-	const evePresenter = createEveSlackPresenter({ ticker });
-	// Reactions are best-effort acknowledgment; never fail the run on them.
-	const reactSafely = (input: { action: "add" | "remove"; emoji: string }) =>
-		react?.(input).catch(() => undefined);
-	// Status starts before any lookups so the thread never sits silent.
-	ticker.thinking();
-	try {
-		const workspaceId = getSlackWorkspaceId(raw);
-		// The Slack history fetch (recentMessages) overlaps the DB lookup.
-		const [installation, recentMessages] = await Promise.all([
-			findSlackInstallationForWorkspace({ workspaceId }),
-			Promise.resolve(
-				typeof recentMessagesInput === "function"
-					? recentMessagesInput()
-					: recentMessagesInput,
-			),
-		]);
-		if (!installation) {
-			logger.warn("Slack installation not found", {
-				event: "leaf.slack_installation_missing",
-			});
-			return;
-		}
-
-		const session = createLeafSessionContext({
-			channelId,
-			provider: installation.provider,
-			providerUserId,
-			threadId,
-			workspaceId,
-		});
-		logger = addLeafContext(rootLogger, {
-			...session.context,
-			agent_run_id: session.agentRunId,
-			org_id: installation.org_id,
-			org_slug: installation.org_slug,
-		});
-		logger.info("Received Slack message", {
-			event: "leaf.slack_message_received",
-			data: {
-				attachment_count: attachments?.length ?? 0,
-				text_length: text.length,
-			},
-		});
-		if (!text.trim() && !attachments?.length) {
-			logger.info("Skipping empty Slack message", {
-				event: "leaf.slack_message_skipped",
-				data: { reason: "empty" },
-			});
-			return;
-		}
-
-		run = registerRun({
-			key: runKey,
-			kind: "message",
-			ownerProviderUserId: providerUserId,
-		});
-		const logAction = (message: string) => {
-			ticker.activity(message);
-		};
-		run.logAction = logAction;
-		const rawFiles = getSlackFilesFromRaw({ raw });
-		const botToken = decrypt(installation.bot_access_token);
-
-		const output = await runSlackAgentTurn({
-			agentRunId: session.agentRunId,
-			attachmentFetchFallback: ({ attachment }) =>
-				fetchSlackAttachmentFallback({
-					attachment,
-					botToken,
-					rawFiles,
-				}),
-			attachments,
-			clientContext,
-			installation,
-			logger,
-			onAction: logAction,
-			onApprovalsSuperseded: (approvals) =>
-				editSupersededApprovalCards({ approvals, logger, target }),
-			onReasoning: evePresenter.onReasoning,
-			onThinking: ticker.thinking,
-			providerUserId,
-			recentMessages,
-			run,
-			text,
-			channelId,
-			threadId,
-		});
-
-		if (output.kind === "stopped") {
-			ticker.stop();
-			const notice =
-				output.reason === "timeout"
-					? RUN_STOPPED_FOR_TIME_MESSAGE
-					: runStoppedByUserNotice(run.stop?.byUserId);
-			await target.post({
-				markdown: [output.text, notice]
-					.filter((part): part is string => Boolean(part?.trim()))
-					.join("\n\n"),
-			});
-			logger.info("Posted stopped run notice", {
-				event: "leaf.slack_run_stopped",
-				data: { stop_reason: output.reason },
-			});
-			return;
-		}
-
-		const outputInstallation =
-			output.kind === "blocked" ? installation : output.installation;
-		const orgId =
-			output.kind === "blocked" ? outputInstallation.org_id : output.org.id;
-
-		if (output.kind === "blocked") {
-			ticker.stop();
-			await target.post({ markdown: output.text });
-			return;
-		}
-		const outputText = output.kind === "empty" ? "" : output.text;
-
-		if (hasQueuedThreadMessage(runKey)) {
-			if (output.kind === "approval") {
-				try {
-					await withdrawAgentTurnApproval({
-						approval: output.approval,
-						auth: {
-							appEnv: output.env,
-							channelId,
-							orgId,
-							provider: outputInstallation.provider,
-							providerUserId,
-							threadId,
-							workspaceId: outputInstallation.workspace_id,
-						},
-						orgId,
-						sessionId: output.sessionId,
-					});
-				} catch (error) {
-					logger.warn("Could not withdraw suspension for queued message", {
-						event: "leaf.eve_queued_withdraw_failed",
-						error,
-					});
-				}
-			}
-			logger.info("Suppressed reply; newer message queued", {
-				event: "leaf.slack_reply_suppressed",
-				data: { had_suspension: output.kind === "approval" },
-			});
-			return;
-		}
-
-		const resolveCatalogDecisionSafely = async () => {
-			try {
-				if (output.kind === "approval") {
-					logAction("Reviewing versioning impact");
-				}
-				return await resolveAgentCatalogDecision({
-					decisionProvided: Boolean(clientContext?.catalogDecision),
-					env: output.env,
-					getToken: () =>
-						getInstallationOAuthAccessToken({
-							installation: outputInstallation,
-							env: output.env,
-							orgId,
-						}),
-					logger,
-					orgId,
-					providerUserId,
-					thread: {
-						channelId,
-						provider: outputInstallation.provider,
-						threadId,
-						workspaceId: outputInstallation.workspace_id,
-					},
-					turn: output,
-				});
-			} catch (error) {
-				logger.warn("Could not evaluate catalog decision redirect", {
-					event: "leaf.eve_catalog_redirect_failed",
-					error,
-				});
-				return undefined;
-			}
-		};
-		const catalogDecision = await resolveCatalogDecisionSafely();
-
-		ticker.stop();
-
-		if (catalogDecision) {
-			if (outputText.trim()) {
-				await target.post({ markdown: outputText });
-			}
-			await target.post(
-				catalogDecisionCard({
-					env: output.env,
-					model: buildCatalogDecisionModel({ plan: catalogDecision.plan }),
-					orgId,
-					plan: catalogDecision.plan,
-				}),
-			);
-			return;
-		}
-
-		if (output.kind === "question") {
-			await target.post(
-				questionCard({
-					env: output.env,
-					options: output.question.options,
-					orgId,
-					prompt: output.question.prompt,
-					requestId: output.question.requestId,
-					sessionId: output.sessionId,
-				}),
-			);
-			return;
-		}
-
-		if (output.kind === "approval") {
-			const postedApproval = await presentApproval({
-				channelId,
-				env: output.env,
-				installation: outputInstallation,
-				logAction,
-				logger,
-				orgId,
-				providerUserId,
-				target,
-				turn: output,
-			});
-			if (postedApproval) return;
-		}
-
-		if (!outputText.trim()) {
-			await target.post({ markdown: `:warning: ${NO_REPLY_MESSAGE}` });
-			logger.warn("Agent produced no reply", {
-				event: "leaf.slack_empty_response",
-				data: {
-					kind: output.kind,
-					run_id: output.sessionId,
-				},
-			});
-			return;
-		}
-
-		await target.post({ markdown: outputText });
-		logger.info("Posted Slack response", {
-			event: "leaf.slack_response_posted",
-			data: {
-				has_text: true,
-			},
-		});
-	} catch (error) {
-		logger.error("[chat] Message failed", error, {
-			event: "leaf.slack_message_failed",
-		});
-		// Stop the status loop before posting, or a pending tick re-renders
-		// "Thinking…" over the cleared status after the error message lands.
-		ticker.stop();
-		await reactSafely({ action: "add", emoji: "x" });
-		await target.post({
-			markdown: `:warning: ${errorNotice(error)}`,
-		});
-	} finally {
-		ticker.stop();
-		await reactSafely({ action: "remove", emoji: "eyes" });
-		if (run) closeRun({ key: run.key, run });
-	}
-};
-
-const handleMessage = async (thread: Thread, message: Message) => {
-	// Never respond to other bots (including a second Autumn app on the same
-	// workspace) — otherwise two bots reply to each other in an infinite loop.
-	if (message.author.isBot === true) {
-		rootLogger.info("Skipping bot-authored Slack message", {
-			event: "leaf.slack_message_skipped",
-			data: { reason: "bot_author" },
-		});
-		return;
-	}
-	const runKey = slackRunKey({
-		channelId: thread.channelId,
-		raw: message.raw,
-		threadId: thread.id,
-	});
-	// Acknowledge instantly, before any lookups — the 👀 lands while the run
-	// (or the queue ahead of it) is still spinning up.
-	thread.adapter.addReaction(thread.id, message.id, "eyes").catch(() => {});
-	await dispatchThreadMessage({
-		hasAttachments: Boolean(message.attachments?.length),
-		providerUserId: message.author.userId,
-		runKey,
-		// Recent messages are fetched when the run actually starts, so a
-		// mutex-queued message still sees the turns that finished before it.
-		runNewMessage: async () =>
-			runAndReply({
-				target: thread,
-				attachments: message.attachments,
-				raw: message.raw,
-				react: async ({ action, emoji }) => {
-					if (action === "add") {
-						await thread.adapter.addReaction(thread.id, message.id, emoji);
-					} else {
-						await thread.adapter.removeReaction(thread.id, message.id, emoji);
-					}
-				},
-				runKey,
-				text: message.text,
-				channelId: thread.channelId,
-				providerUserId: message.author.userId,
-				threadId: thread.id,
-				// A thunk on purpose: fetched at run start (after any queue wait),
-				// in parallel with the installation lookup.
-				recentMessages: () => getRecentMessages(thread, message),
-			}),
-		text: message.text,
-	});
-};
-
-bot.onDirectMessage(handleMessage);
-
+bot.onDirectMessage(handleSlackMessage);
 bot.onNewMention(async (thread, message) => {
 	await thread.subscribe();
-	await handleMessage(thread, message);
+	await handleSlackMessage(thread, message);
 });
-
-bot.onSubscribedMessage(handleMessage);
-
-bot.onSlashCommand(async (event) => {
-	const runKey = slackRunKey({
-		channelId: event.channel.id,
-		raw: event.raw,
-		threadId: event.channel.id,
-	});
-	await dispatchThreadMessage({
-		hasAttachments: false,
-		providerUserId: event.user.userId,
-		runKey,
-		runNewMessage: () =>
-			runAndReply({
-				target: event.channel,
-				raw: event.raw,
-				runKey,
-				text: event.text || event.command,
-				channelId: event.channel.id,
-				providerUserId: event.user.userId,
-				threadId: event.channel.id,
-			}),
-		text: event.text || event.command,
-	});
-});
-
+bot.onSubscribedMessage(handleSlackMessage);
+bot.onSlashCommand(handleSlackSlashCommand);
 bot.onAction(
 	["approve_billing_action", "cancel_billing_action"],
 	handleApprovalAction,
 );
-
-const parseButtonPayload = <T>(value?: string): T | null => {
-	if (!value) return null;
-	try {
-		return JSON.parse(value) as T;
-	} catch {
-		return null;
-	}
-};
-
-// A clicked answer chip on a parked ask_question: answer eve structurally,
-// then relay whatever the resumed turn produced (text, another question, or a
-// chained gated write).
-bot.onAction(indexedActionIds(ANSWER_QUESTION_ACTION), async (event) => {
-	const payload = parseButtonPayload<QuestionButtonPayload>(event.value);
-	if (!payload) return;
-	const workspaceId = getSlackWorkspaceId(event.raw);
-	const installation = await findSlackInstallationForWorkspace({ workspaceId });
-	if (!installation) return;
-	const providerUserId = event.user.userId;
-
-	// Collapse the buttons immediately so double-clicks read as answered.
-	try {
-		await event.adapter.editMessage?.(
-			event.threadId,
-			event.messageId,
-			questionAnsweredCard({
-				actorId: providerUserId,
-				answerLabel: payload.l,
-				prompt: payload.q,
-			}),
-		);
-	} catch {
-		// Cosmetic; the structural answer below is what matters.
-	}
-	try {
-		await event.thread?.startTyping("Working on it...");
-	} catch {
-		// Cosmetic.
-	}
-
-	try {
-		const result = await answerAgentQuestion({
-			auth: {
-				appEnv: payload.e as AppEnv,
-				channelId: event.thread?.channelId ?? event.threadId,
-				orgId: payload.g,
-				provider: installation.provider,
-				providerUserId,
-				threadId: event.threadId,
-				workspaceId: installation.workspace_id,
-			},
-			optionId: payload.o,
-			orgId: payload.g,
-			requestId: payload.r,
-			sessionId: payload.s,
-		});
-		if ("error" in result) {
-			await event.thread?.post({
-				markdown: questionAnswerFailedWithDetail(result.message),
-			});
-			return;
-		}
-		if (result.text.trim()) {
-			await event.thread?.post({ markdown: result.text });
-		}
-		if (result.question) {
-			await event.thread?.post(
-				questionCard({
-					env: payload.e as AppEnv,
-					options: result.question.options,
-					orgId: payload.g,
-					prompt: result.question.prompt,
-					requestId: result.question.requestId,
-					sessionId: result.sessionId,
-				}),
-			);
-		}
-		if (result.chainedApprovalId && event.thread) {
-			const chained = await chatApprovalRepo.get({
-				approvalId: result.chainedApprovalId,
-				db,
-			});
-			if (chained) {
-				await postApprovalCardForRow({
-					approval: chained,
-					target: event.thread,
-				});
-			}
-		}
-	} catch (error) {
-		rootLogger.error("[chat] Question answer failed", error, {
-			event: "leaf.eve_question_answer_failed",
-		});
-		await event.thread?.post({ markdown: QUESTION_ANSWER_FAILED_MESSAGE });
-	}
-});
-
-// A clicked catalog decision: dispatch a new run carrying the decision as
-// one-turn clientContext, exactly like the dashboard's decision card.
-bot.onAction(indexedActionIds(CATALOG_DECISION_ACTION), async (event) => {
-	const payload = parseButtonPayload<CatalogDecisionButtonPayload>(event.value);
-	if (!(payload && event.thread)) return;
-	const thread = event.thread;
-	const providerUserId = event.user.userId;
-	try {
-		await event.adapter.editMessage?.(
-			event.threadId,
-			event.messageId,
-			catalogDecisionSubmittedCard({
-				actorId: providerUserId,
-				choiceLabel: payload.l,
-				planName: payload.p,
-			}),
-		);
-	} catch {
-		// Cosmetic; the dispatched run is what matters.
-	}
-
-	const decision = {
-		migrationDraft: payload.m === 1,
-		planId: payload.p,
-		propagateVariantIds: payload.pv,
-		versioning: payload.v,
-	};
-	const text = `I chose "${payload.l}" for ${payload.p} on the catalog decision card. Apply the change with these decisions.`;
-	const runKey = slackRunKey({
-		channelId: thread.channelId,
-		raw: event.raw,
-		threadId: event.threadId,
-	});
-	await dispatchThreadMessage({
-		hasAttachments: false,
-		providerUserId,
-		runKey,
-		runNewMessage: () =>
-			runAndReply({
-				// ActionEvent threads carry an unknown state generic; posting works
-				// identically to message-handler threads.
-				target: thread as unknown as ReplyTarget,
-				clientContext: { catalogDecision: decision },
-				raw: event.raw,
-				runKey,
-				text,
-				channelId: thread.channelId,
-				providerUserId,
-				threadId: event.threadId,
-			}),
-		text,
-	});
-});
-
+bot.onAction(questionAnswerActionIds, handleSlackQuestionAnswer);
+bot.onAction(catalogDecisionActionIds, handleSlackCatalogDecision);
 bot.onAction(["view_approval_payload"], handleViewPayloadAction);
-
 bot.onAction(["stop_agent_run"], handleStopAction);
-
 bot.registerSingleton();
